Move GitHub integration into dedicated crate
Moves the initialization and authentication flow for the GitHub integration from but-api into a new dedicated but-github crate. This includes creating the new but-github crate, refactoring OAuth and authentication related code to the new location, and updating dependencies across workspace crates to use but-github accordingly.
