+++ crates/but-api/Cargo.toml
@@ -20,6 +20,7 @@ tracing.workspace = true
 tauri = { version = "^2.7.0", features = ["unstable"] } # For the #[tauri::command(async)] macro only
 
 but-settings.workspace = true
+but-github.workspace = true
 gitbutler-user.workspace = true
 gitbutler-project.workspace = true
 gitbutler-secret.workspace = true
