+++ crates/but-github/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "but-github"
+version = "0.0.0"
+edition = "2024"
+repository = "https://github.com/gitbutlerapp/gitbutler"
+license-file = "../../LICENSE.md"
+description = "A GitButler CLI tool"
+authors = ["GitButler <gitbutler@gitbutler.com>"]
+readme = "../../README.md"
+publish = false
+rust-version = "1.89"
+
+[dependencies]
+serde.workspace = true
+but-settings.workspace = true
+anyhow.workspace = true
+serde_json = "1.0.145"
+reqwest = { version = "0.12", features = ["json"] }
