+++ crates/gitbutler-tauri/Cargo.toml
@@ -68,6 +68,7 @@ but-action.workspace = true
 but-bot.workspace = true
 but-api.workspace = true
 but-claude.workspace = true
+but-github.workspace = true
 open = "5"
 url = "2.5.4"
 but-path.workspace = true
