+++ crates/gitbutler-tauri/src/github.rs
@@ -1,7 +1,8 @@
 use but_api::{
-    commands::github::{self, Verification},
+    commands::github::{self},
     NoParams,
 };
+use but_github::{CheckAuthStatusParams, Verification};
 use tracing::instrument;
 
 use but_api::error::Error;
@@ -15,5 +16,5 @@ pub async fn init_device_oauth() -> Result<Verification, Error> {
 #[tauri::command(async)]
 #[instrument(err(Debug))]
 pub async fn check_auth_status(device_code: String) -> Result<String, Error> {
-    github::check_auth_status(github::CheckAuthStatusParams { device_code }).await
+    github::check_auth_status(CheckAuthStatusParams { device_code }).await
 }
